+++ src/main/java/com/afterburner/AfterburnerBackendApplication.java
@@ -1,12 +1,9 @@
 package com.afterburner;
 
-import com.afterburner.oauth.config.AppProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
-@EnableConfigurationProperties(AppProperties.class)
 public class AfterburnerBackendApplication {
 
 	public static void main(String[] args) {
