--- src/main/java/com/afterburner/oauth/controller/UserController.java
@@ -1,26 +1,0 @@
-package com.afterburner.oauth.controller;
-
-
-import com.afterburner.oauth.exception.ResourceNotFoundException;
-import com.afterburner.oauth.model.User;
-import com.afterburner.oauth.repository.UserRepository;
-import com.afterburner.oauth.security.CurrentUser;
-import com.afterburner.oauth.security.UserPrincipal;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @GetMapping("/user/me")
-    @PreAuthorize("hasRole('USER')")
-    public User getCurrentUser(@CurrentUser UserPrincipal userPrincipal) {
-        return userRepository.findById(userPrincipal.getId())
-                .orElseThrow(() -> new ResourceNotFoundException("User", "id", userPrincipal.getId()));
-    }
-}
