--- src/main/java/com/afterburner/oauth/exception/OAuth2AuthenticationProcessingException.java
@@ -1,13 +1,0 @@
-package com.afterburner.oauth.exception;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class OAuth2AuthenticationProcessingException extends AuthenticationException {
-    public OAuth2AuthenticationProcessingException(String msg, Throwable t) {
-        super(msg, t);
-    }
-
-    public OAuth2AuthenticationProcessingException(String msg) {
-        super(msg);
-    }
-}
