--- src/main/java/com/afterburner/oauth/model/AuthProvider.java
@@ -1,8 +1,0 @@
-package com.afterburner.oauth.model;
-
-public enum  AuthProvider {
-    local,
-    facebook,
-    google,
-    github
-}
