--- src/main/java/com/afterburner/oauth/payload/ApiResponse.java
@@ -1,27 +1,0 @@
-package com.afterburner.oauth.payload;
-
-public class ApiResponse {
-    private boolean success;
-    private String message;
-
-    public ApiResponse(boolean success, String message) {
-        this.success = success;
-        this.message = message;
-    }
-
-    public boolean isSuccess() {
-        return success;
-    }
-
-    public void setSuccess(boolean success) {
-        this.success = success;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
