--- src/main/java/com/afterburner/oauth/payload/AuthResponse.java
@@ -1,26 +1,0 @@
-package com.afterburner.oauth.payload;
-
-public class AuthResponse {
-    private String accessToken;
-    private String tokenType = "Bearer";
-
-    public AuthResponse(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getTokenType() {
-        return tokenType;
-    }
-
-    public void setTokenType(String tokenType) {
-        this.tokenType = tokenType;
-    }
-}
