--- src/main/java/com/afterburner/oauth/payload/LoginRequest.java
@@ -1,33 +1,0 @@
-package com.afterburner.oauth.payload;
-
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-/**
- * Created by rajeevkumarsingh on 02/08/17.
- */
-public class LoginRequest {
-    @NotBlank
-    @Email
-    private String email;
-
-    @NotBlank
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
