--- src/main/java/com/afterburner/oauth/repository/UserRepository.java
@@ -1,18 +1,0 @@
-package com.afterburner.oauth.repository;
-
-
-import com.afterburner.oauth.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    Optional<User> findByEmail(String email);
-
-    Boolean existsByEmail(String email);
-
-}
