--- src/main/java/com/afterburner/oauth/security/CurrentUser.java
@@ -1,13 +1,0 @@
-package com.afterburner.oauth.security;
-
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.PARAMETER, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@AuthenticationPrincipal
-public @interface CurrentUser {
-
-}
