--- src/main/java/com/afterburner/oauth/security/CustomUserDetailsService.java
@@ -1,42 +1,0 @@
-package com.afterburner.oauth.security;
-
-import com.afterburner.oauth.exception.ResourceNotFoundException;
-import com.afterburner.oauth.model.User;
-import com.afterburner.oauth.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * Created by rajeevkumarsingh on 02/08/17.
- */
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    UserRepository userRepository;
-
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String email)
-            throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() ->
-                        new UsernameNotFoundException("User not found with email : " + email)
-        );
-
-        return UserPrincipal.create(user);
-    }
-
-    @Transactional
-    public UserDetails loadUserById(Integer id) {  // Integer 타입으로 수정
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User", "id", id));
-
-        return UserPrincipal.create(user);
-    }
-}
