--- src/main/java/com/afterburner/oauth/security/oauth2/user/GoogleOAuth2UserInfo.java
@@ -1,30 +1,0 @@
-package com.afterburner.oauth.security.oauth2.user;
-
-import java.util.Map;
-
-public class GoogleOAuth2UserInfo extends OAuth2UserInfo {
-
-    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getName() {
-        return (String) attributes.get("name");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getImageUrl() {
-        return (String) attributes.get("picture");
-    }
-}
