--- src/main/java/com/afterburner/oauth/security/oauth2/user/OAuth2UserInfo.java
@@ -1,23 +1,0 @@
-package com.afterburner.oauth.security.oauth2.user;
-
-import java.util.Map;
-
-public abstract class OAuth2UserInfo {
-    protected Map<String, Object> attributes;
-
-    public OAuth2UserInfo(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    public abstract String getId();
-
-    public abstract String getName();
-
-    public abstract String getEmail();
-
-    public abstract String getImageUrl();
-}
