--- src/main/java/com/afterburner/oauth/security/oauth2/user/OAuth2UserInfoFactory.java
@@ -1,22 +1,0 @@
-package com.afterburner.oauth.security.oauth2.user;
-
-
-import com.afterburner.oauth.exception.OAuth2AuthenticationProcessingException;
-import com.afterburner.oauth.model.AuthProvider;
-
-import java.util.Map;
-
-public class OAuth2UserInfoFactory {
-
-    public static OAuth2UserInfo getOAuth2UserInfo(String registrationId, Map<String, Object> attributes) {
-        if(registrationId.equalsIgnoreCase(AuthProvider.google.toString())) {
-            return new GoogleOAuth2UserInfo(attributes);
-        } else if (registrationId.equalsIgnoreCase(AuthProvider.facebook.toString())) {
-            return new FacebookOAuth2UserInfo(attributes);
-        } else if (registrationId.equalsIgnoreCase(AuthProvider.github.toString())) {
-            return new GithubOAuth2UserInfo(attributes);
-        } else {
-            throw new OAuth2AuthenticationProcessingException("Sorry! Login with " + registrationId + " is not supported yet.");
-        }
-    }
-}
